consistent channel -> socket messages
Previously each serializer determined the message format which introduced a strong coupling between the serializer and the socket process. The shape of messages arriving to the socket process depended on the serializer (which is transport related).

Using consistent message format allows us to handle messages to the socket process in the same way, regardless of the encoder/fastlaner/transport which is used.